creating a new file within nipype.testing that contains various pytest fixtures; moving several different versions of create_files_in_directory fixture to nipype.testing.fixtures
